examples/04_web_api_fastapi/requirements.txt
lunita-mini
python-dotenv
fastapi
uvicorn
